Refactor tests
Add comments, move tests, use defaultIntercepts, add new cypress commands
