+++ packages/fyllut/cypress/e2e/form/list-forms.ts
@@ -1,3 +1,7 @@
+/*
+ * Tests the page listing out all forms
+ */
+
 describe('Skjemaliste', () => {
   beforeEach(() => {
     cy.defaultIntercepts();
