+++ packages/fyllut/cypress/e2e/form/prefill-data.cy.ts
@@ -1,3 +1,7 @@
+/*
+ * Tests that the fields that have prefillKey set in the form definition will be prefilled with data
+ */
+
 describe('Prefill data', () => {
   before(() => {
     cy.configMocksServer();
@@ -6,7 +10,7 @@ describe('Prefill data', () => {
   beforeEach(() => {
     cy.defaultIntercepts();
     cy.defaultInterceptsMellomlagring();
-    cy.defaultInterceptsPrefillData();
+    cy.defaultInterceptsExternal();
     cy.mocksRestoreRouteVariants();
   });
 
@@ -17,7 +21,7 @@ describe('Prefill data', () => {
   describe('new application', () => {
     it('should prefill data for new application on the first page (name)', () => {
       cy.visit('/fyllut/testprefilldata?sub=digital');
-      cy.wait('@getTestFormPrefillData');
+      cy.wait('@getForm');
       cy.clickStart();
       cy.wait('@getPrefillData');
       cy.wait('@createMellomlagring');
@@ -29,7 +33,7 @@ describe('Prefill data', () => {
 
     it('should prefill data for new application on the second page (name)', () => {
       cy.visit('/fyllut/testprefilldata?sub=digital');
-      cy.wait('@getTestFormPrefillData');
+      cy.wait('@getForm');
       cy.clickStart();
       cy.wait('@getPrefillData');
       cy.wait('@createMellomlagring');
@@ -42,7 +46,7 @@ describe('Prefill data', () => {
 
     it('should not prefill data for new application if submissionMethod is paper', () => {
       cy.visit('/fyllut/testprefilldata?sub=paper');
-      cy.wait('@getTestFormPrefillData');
+      cy.wait('@getForm');
       cy.clickStart();
 
       // Should not make a request to get prefill data
@@ -61,7 +65,7 @@ describe('Prefill data', () => {
       cy.mocksUseRouteVariant('get-soknad:success-prefill-data');
 
       cy.visit('/fyllut/testprefilldata/side1?sub=digital&innsendingsId=d2f41ebc-ba98-4fc5-a195-29b098bf50a7');
-      cy.wait('@getTestFormPrefillData');
+      cy.wait('@getForm');
       cy.wait('@getPrefillData');
 
       cy.findByRole('heading', { name: 'Side 1' }).should('exist');
@@ -77,7 +81,7 @@ describe('Prefill data', () => {
       cy.mocksUseRouteVariant('get-soknad:success-prefill-data');
 
       cy.visit('/fyllut/testprefilldata/side1?sub=digital&innsendingsId=d2f41ebc-ba98-4fc5-a195-29b098bf50a7');
-      cy.wait('@getTestFormPrefillData');
+      cy.wait('@getForm');
       cy.wait('@getPrefillData');
       cy.clickSaveAndContinue();
